refactor: CartItem, product 타입 분리
