--- src/types/CartItems.ts
@@ -1,1 +1,0 @@
-export type CartItems = { [itemID: number]: { quantity: number } };
